Muda hps para floats
Assim evita erros de arredondamento, por exemplo se algo for causar de dano 12% do hp de outra

+++ jogo42-cb/desenhar.c
@@ -98,7 +98,7 @@ static void DesenharHpJog(const GameState* gs)
 {
     const int POS_X = 10, POS_Y = 10, TAM_FONTE = 20;
 
-    DrawText(FormatText("HP: %d", gs->jog.hp),
+    DrawText(FormatText("HP: %d", (int)gs->jog.hp),
              POS_X, POS_Y, TAM_FONTE, WHITE);
 }
